+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 ### Updates:
 
-* [PHP `7.3.3` `7.2.16` `7.1.27`](http://www.php.net/ChangeLog-7.php#7.3.3)
+* [PHP `7.3.3` `7.2.16` `7.1.27`](https://www.php.net/ChangeLog-7.php#7.3.3)
 * [Redis `5.0.4`](https://raw.githubusercontent.com/antirez/redis/5.0/00-RELEASENOTES)
-* [NGINX `1.15.10`](http://nginx.org/en/CHANGES)
+* [NGINX `1.15.10`](https://nginx.org/en/CHANGES)
+* [HTTPD `2.4.39`](https://www.apache.org/dist/httpd/CHANGES_2.4.39)