+++ CHANGELOG.md
@@ -13,3 +13,4 @@
 ### Updates:
 
 * [PHP `7.3.4` `7.2.17` `7.1.28`](https://www.php.net/ChangeLog-7.php#7.3.4)
+* [NGINX `1.15.12`](https://nginx.org/en/CHANGES)